+++ src/main/java/com/nikolastojanovic/rbtticketingsystem/domain/model/Order.java
@@ -4,6 +4,7 @@ import com.nikolastojanovic.rbtticketingsystem.domain.model.enums.OrderMethod;
 import com.nikolastojanovic.rbtticketingsystem.domain.model.enums.OrderStatus;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Builder
@@ -12,7 +13,7 @@ public record Order(
         Long userId,
         Long eventId,
         Integer ticketCount,
-        Double totalAmount,
+        BigDecimal totalAmount,
         OrderStatus status,
         OrderMethod orderMethod,
         ZonedDateTime createdAt,
